Update README.md
This program is based on Linear Regression Algorithm of Machine Learning where the data consisting of 'Movie_Revenue' and 'Production_Cost' is fed into the program as learning data and the result is delivered for the test data the user enters.

The user can enter the test budget cost in USD and obtain the cost in USD the movie will supposedly earn while premiering. 

For running this program, it is mandatory to have Jupyter_Notebook pre-installed on the system or the online version also works.

To enter your own raw data excel sheet, simply rename the file as 'cost_revenue_clean' and rename the column names as 'production_budget_usd' and 'worldwide_gross_usd' respectively.